fix: 🚑️ moved `AuthProvider` from `layout.tsx` to `page.tsx`
caused continues invoking of `/api/auth/session`

Related: #166
