+++ src/app/admin/page.tsx
@@ -1,3 +1,11 @@
-export default function AdminView() {
-  return <p>Hello Admin</p>;
+import AuthProvider from '@/components/AuthProvider';
+
+export default function AdminPage() {
+  return (
+    <AuthProvider>
+      <main>
+        <p>Hello Admin</p>
+      </main>
+    </AuthProvider>
+  );
 }
